requirements.txt
fastapi
uvicorn
pydantic
psutil
pyautogui
pillow
openai
openenv-core
python-xlib